% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipFileOutput.R
\name{downloadBigUI}
\alias{downloadBigUI}
\title{Shiny UI module for data download}
\usage{
downloadBigUI(id, label)
}
\arguments{
\item{id}{Namespace identifier}

\item{label}{Download button label}
}
\value{
Download of Sample resources
}
\description{
Designed to handle large frame files by a separated compilation and download process.
}
