% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameCheck.R
\name{frameCheck}
\alias{frameCheck}
\title{CHECK EA size}
\usage{
frameCheck(data = NULL, STRATUM = NULL, PSU = "EACODE2", n_hh = 12,
  COUNT = "count")
}
\arguments{
\item{data}{frame dataset}

\item{STRATUM}{stratum identifier}

\item{PSU}{psu identifier}

\item{n_hh}{minium psu size}

\item{COUNT}{varialbe containing population count}
}
\description{
VERSION 2.0.0
}
