% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingCube_selection.R
\name{samplingCubeInput}
\alias{samplingCubeInput}
\title{Shiny UI module to select variables for balanced sampling}
\usage{
samplingCubeInput(id)
}
\arguments{
\item{id}{Namespace identifier}
}
\value{
Sample Size
}
\description{
Shiny UI module to select variables for balanced sampling
}
