% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingCube_start.R
\name{samplingCubeSRV}
\alias{samplingCubeSRV}
\title{Shiny server module for simulation start}
\usage{
samplingCubeSRV(input, output, session, dataset, sampleSize,
  balancedVariable, target_var, nsample = 1000, seed)
}
\arguments{
\item{input}{standard shiny input}

\item{output}{standard shiny output}

\item{session}{standard shiny session object}

\item{dataset}{frame dataset used for the stratification}

\item{sampleSize}{sample size}

\item{balancedVariable}{set of variables to balance on}

\item{target_var}{the variable of interest}

\item{nsample}{Number of Simulation runs (recommended is 1000)}

\item{seed}{Set seed for MC with doRNG}
}
\value{
Data frame containing the final sample
}
\description{
Shiny server module for simulation start
}
