% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingCube_start.R
\name{samplingCubeUI}
\alias{samplingCubeUI}
\title{Shiny UI module to start balanced sampling}
\usage{
samplingCubeUI(id)
}
\arguments{
\item{id}{Namespace identifier}
}
\value{
rhandsontable with SamplingStrata input parameters
}
\description{
Shiny UI module to start balanced sampling
}
