% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingStrata_selection.R
\name{samplingStrataInput}
\alias{samplingStrataInput}
\title{Shiny UI module for SamplingStrata variable selection}
\usage{
samplingStrataInput(id)
}
\arguments{
\item{id}{Namespace identifier}
}
\value{
file with data.table of uploaded data
}
\description{
This module creates the inputs and also updates them after dataset is
available. Due to the construction of this module, the inputs are availble in the
global environment under input$[id-ns(inputs)]. This means to reference i.e to
the selectizeInput ns("strat_var_cont) with the hypotetical id "stratInputs" you have
to use input$`stratInputs-strat_var_cont`.
}
