% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingStrata_params.R
\name{samplingStrataInput_para}
\alias{samplingStrataInput_para}
\title{Shiny UI module to create a rhandsontable for SamplingStrata inputs}
\usage{
samplingStrataInput_para(id)
}
\arguments{
\item{id}{Namespace identifier}
}
\value{
rhandsontable with SamplingStrata input parameters
}
\description{
Shiny UI module to create a rhandsontable for SamplingStrata inputs
}
