% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingStrata_params.R
\name{samplingStrataOutput_para}
\alias{samplingStrataOutput_para}
\title{Shiny server module to create a rhandsontable for SamplingStrata inputs}
\usage{
samplingStrataOutput_para(input, output, session, dataset, target_var,
  domain_var)
}
\arguments{
\item{input}{standard shiny input}

\item{output}{standard shiny output}

\item{session}{standard shiny session object}

\item{dataset}{frame dataset used for the stratification}

\item{target_var}{the selected target variables}

\item{domain_var}{the selected domain variables}
}
\description{
Shiny server module to create a rhandsontable for SamplingStrata inputs
}
