% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingStrata_start.R
\name{samplingStrataSRV}
\alias{samplingStrataSRV}
\title{Shiny server module for simulation start}
\usage{
samplingStrataSRV(input, output, session, dataset, domain_var, target_var,
  strat_var_cat, strat_var_cont, start_strat, frame_CV_in, nBins, seed,
  minStrat)
}
\arguments{
\item{input}{standard shiny input}

\item{output}{standard shiny output}

\item{session}{standard shiny session object}

\item{dataset}{frame dataset used for the stratification}

\item{domain_var}{the selected domain variables}

\item{target_var}{the selected target variables}

\item{strat_var_cat}{the selected categorical stratification variables}

\item{strat_var_cont}{the selected continous stratification variables (require bin to be set)}

\item{start_strat}{simulation start button}

\item{frame_CV_in}{dataframe of CVs by domain}

\item{nBins}{number of bins for the k-means clustering implemented by \code{SamplingStrata::var.bin()} for continous stratification variables}

\item{seed}{the random seed required for the optimization to be reproducable}

\item{minStrat}{this is the minimum number of units required per stratum}
}
\value{
Data frame containing the final sample
}
\description{
Shiny server module for simulation start
}
