% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingStrata_start.R
\name{samplingStrataUI}
\alias{samplingStrataUI}
\title{Shiny UI module for simulation start}
\usage{
samplingStrataUI(id)
}
\arguments{
\item{id}{Namespace identifier}
}
\value{
CV and Relative Bias plot
}
\description{
Shiny UI module for simulation start
}
