% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationDiagnostics.R
\name{simuDiagnosticSRV}
\alias{simuDiagnosticSRV}
\title{Shiny server module for simulation diagnostic results}
\usage{
simuDiagnosticSRV(input, output, session, dataset)
}
\arguments{
\item{input}{standard shiny input}

\item{output}{standard shiny output}

\item{session}{standard shiny session object}

\item{dataset}{is the dataset from the simulation output, must be a \emph{data.table}}
}
\description{
This module creates the outputs for the results from a sampling
simulation. The frame must contain at least the columns \emph{Mean, SE, CV, bias}. In cases of domain
or stratified sampling, a 5th column may be included named \emph{Domain}.
}
