% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationDiagnostics.R
\name{simuDiagnosticUI}
\alias{simuDiagnosticUI}
\title{Shiny UI module for simulation diagnostic results}
\usage{
simuDiagnosticUI(id)
}
\arguments{
\item{id}{Namespace identifier}
}
\value{
rhandsontable with SamplingStrata input parameters
}
\description{
Shiny UI module for simulation diagnostic results
}
