% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveySolutionsCOVID19tools.R
\docType{package}
\name{SurveySolutionsCOVID19tools}
\alias{SurveySolutionsCOVID19tools}
\alias{SurveySolutionsCOVID19tools-package}
\title{Survey Solutons COVID19 survey toolbox}
\description{
This package contains a set of tools for the effecient implementation of Computer Assisted
Telephone Interview (CATI) surveys. Each of the tools contains the corresponding GUI,
either written in vanilla shiny or in rmarkdown's flexdashboard. This is only a first
version, and frequent updates will follow during the weeks to come, so make sure to
check by.

Currently implemented are:

\itemize{
   \item Sampling
      \itemize{
         \item \href{https://github.com/barcaroli/SamplingStrata}{SamplingStrata}
         \item \href{https://github.com/cran/sampling}{cubeSample (package sampling)}
       }
   \item Quality Control
   \item Analysis and Visualization
}
}
\examples{
\dontrun{
suso_covid19_samplingApp()
}

}
