% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startApp.R
\name{suso_covid19_samplingApp}
\alias{suso_covid19_samplingApp}
\title{Start Sampling Application}
\usage{
suso_covid19_samplingApp(launch.browser = T, quiet = T)
}
\arguments{
\item{launch.browser}{if TRUE starts in your default browser, otherwise in your IDE viewer}

\item{quiet}{if TRUE start-up messages are not printed}
}
\description{
Running this function starts the
the COVID19 CATI Surveys
Sampling Application for household survey sub-sample
in your browser
}
