% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipFileInput.R
\name{zipFile}
\alias{zipFile}
\title{Shiny server module for upload of text (csv, tab) file}
\usage{
zipFile(input, output, session, sep = ",")
}
\arguments{
\item{input}{standard shiny input}

\item{output}{standard shiny output}

\item{session}{standard shiny session object}

\item{sep}{column separator, one of , ; or "\\t"}
}
\value{
returns datatable with upload data by using data.table's fread function
}
\description{
Shiny server module for upload of text (csv, tab) file
}
