% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipFileInput.R
\name{zipFileInput}
\alias{zipFileInput}
\title{Shiny UI module for upload of text (csv, tab) file}
\usage{
zipFileInput(id, label, accept)
}
\arguments{
\item{id}{Namespace identifier}

\item{label}{File input label}

\item{accept}{Mime type (i.e. must be one of)}
}
\value{
file input
}
\description{
Shiny UI module for upload of text (csv, tab) file
}
